Initial refactor to use React
