+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/index.js":"1","/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/App.js":"2","/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/Game.js":"3","/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/Player.js":"4","/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/GamePiece.js":"5"},{"size":222,"mtime":1611871633643,"results":"6","hashOfConfig":"7"},{"size":2483,"mtime":1611872489542,"results":"8","hashOfConfig":"7"},{"size":263,"mtime":1611871917593,"results":"9","hashOfConfig":"7"},{"size":242,"mtime":1611871968789,"results":"10","hashOfConfig":"7"},{"size":133,"mtime":1611871795462,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"1k22en5",{"filePath":"15","messages":"16","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"17","messages":"18","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"19","usedDeprecatedRules":"14"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"22"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/index.js",[],["25","26"],"/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/App.js",["27","28"],"/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/Game.js",["29"],"import Player from \"./Player\";\n\nclass Game {\n  constructor(player1, player2, pongBall, gameCode) {\n    this.player1 = player1;\n    this.player2 = player2;\n    this.pongBall = pongBall;\n    this.round = 1;\n    this.gameCode = gameCode;\n  }\n}\n\nexport default Game;\n","/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/Player.js",["30"],"import GamePiece from \"./GamePiece\";\n\nclass Player {\n  constructor(username, id, paddle) {\n    this.username = username;\n    this.id = id;\n    this.score = 0;\n    this.paddle = paddle;\n  }\n\n  // movePaddle(event) {}\n}\n\nexport default Player;\n","/Users/lopez.4.joseph/turing/fun_projects/video_pong/src/GamePiece.js",[],{"ruleId":"31","replacedBy":"32"},{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","severity":1,"message":"36","line":2,"column":8,"nodeType":"37","messageId":"38","endLine":2,"endColumn":13},{"ruleId":"35","severity":1,"message":"39","line":2,"column":17,"nodeType":"37","messageId":"38","endLine":2,"endColumn":25},{"ruleId":"35","severity":1,"message":"40","line":1,"column":8,"nodeType":"37","messageId":"38","endLine":1,"endColumn":14},{"ruleId":"35","severity":1,"message":"41","line":1,"column":8,"nodeType":"37","messageId":"38","endLine":1,"endColumn":17},"no-native-reassign",["42"],"no-negated-in-lhs",["43"],"no-unused-vars","'React' is defined but never used.","Identifier","unusedVar","'useState' is defined but never used.","'Player' is defined but never used.","'GamePiece' is defined but never used.","no-global-assign","no-unsafe-negation"]
